requirements.txt
fastapi
uvicorn
pdfminer.six
pandas
openpyxl
python-multipart
slowapi
streamlit
requests